Submission COntroller: Handled all common functions inside abstract class
