+++ src/main/java/com/projects/onlinejudge/helpers/IRunner.java
@@ -3,6 +3,72 @@ package com.projects.onlinejudge.helpers;
 import com.projects.onlinejudge.dto.RunRequest;
 import com.projects.onlinejudge.dto.RunResponse;
 
-public interface IRunner {
+import java.io.*;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.stream.Stream;
+
+public abstract class IRunner {
     public abstract RunResponse runTests(RunRequest attribute);
+    public String compare_one_test(String userOutput, String actualOutput ) {
+
+        String userText = removeSpace(userOutput);
+        String actualText = removeSpace(actualOutput);
+        if(userText.equals(actualText)){
+            return "Passed";
+        }
+        else return "Failed";
+        // only sample
+    }
+    public String removeSpace( String file) {
+        StringBuilder contentBuilder = new StringBuilder();
+        try (Stream<String> stream = Files.lines( Paths.get(file), StandardCharsets.UTF_8))
+        {
+            stream.forEach(s -> contentBuilder.append(s));
+        }
+        catch (IOException e)
+        {
+            e.printStackTrace();
+        }
+        return contentBuilder.toString();
+    }
+
+    // Printing the terminal output after code execution
+    public static void printResults(Process process) throws IOException {
+        BufferedReader reader = new BufferedReader(new InputStreamReader(process.getInputStream()));
+        String line = "";
+        while ((line = reader.readLine()) != null) {
+            System.out.println(line);
+        }
+    }
+
+    // This filter will only include files starting  with input
+    FilenameFilter filter = new FilenameFilter() {
+        @Override
+        public boolean accept(File f, String name) {
+            return name.startsWith("input");
+        }
+    };
+
+    public static void copyContent(File a, File b) throws Exception
+    {
+        FileInputStream in = new FileInputStream(a);
+        FileOutputStream out = new FileOutputStream(b);
+        try {
+            int n;
+            while ((n = in.read()) != -1) {
+                out.write(n);
+            }
+        }
+        finally {
+            if (in != null) {
+                in.close();
+            }
+            if (out != null) {
+                out.close();
+            }
+        }
+        System.out.println("File Copied");
+    }
 }
