+++ src/main/java/com/projects/onlinejudge/helpers/JavaRunner.java
@@ -3,7 +3,7 @@ package com.projects.onlinejudge.helpers;
 import com.projects.onlinejudge.dto.RunRequest;
 import com.projects.onlinejudge.dto.RunResponse;
 
-public class JavaRunner implements IRunner{
+public class JavaRunner extends IRunner{
     @Override
     public RunResponse runTests(RunRequest attribute) {
         return null;
