+++ src/main/java/com/projects/onlinejudge/helpers/PythonRunner.java
@@ -3,7 +3,7 @@ package com.projects.onlinejudge.helpers;
 import com.projects.onlinejudge.dto.RunRequest;
 import com.projects.onlinejudge.dto.RunResponse;
 
-public class PythonRunner implements IRunner{
+public class PythonRunner extends IRunner{
     @Override
     public RunResponse runTests(RunRequest attribute) {
         return null;
